Added module function, make_CherenkovPhoton_array, to make full 3D array.
Changed old make_CherernkovPhoton_array to make_CherenkovPhoton_list.
Added function make_CherenkovPhoton_array which has all defaults values,
making an array of t, delta, theta of the Cherenkov Angular distribution.
This function returns a 3D ndarray, and the arrays of t, delta, and theta
values which are meant to be stored together in a npz file.