Added a file defining a class ShowerCherenkov.
ShowerCherenkov, when intitiated, has Gaisser-Hillas parameters for
a shower, and a shower geometry (angles and impact point). It stores
a copy of Atmosphere and CherenkovPhotonArray. It makes an array of
heights, points along the shower, and calculates the slant_depth for
each point. It also stores the time (impact is t=0) the shower was
at the point, the shower stage at each point and the atmospheric delta
at each point. It calculates the Cherernkov shower size, the Cherernkov
yield and a full Cherenkov Angular distribution at each point. Various
instance methods aid these calculations.