Added a class to handle the full CherenkovPhoton array.
Define a class, ChererkovPhotonArray. When instantiated (with an
npz file filename) it contains the 3D array of Cherernkov angular
distributions at t and atmospheric delta. It also has the values
of t, delta and theta that are the values along these planes of
the 3D array. The objects have methds to return a full Cherenkov
angular distribution at a given t, delta values by interpolating
(linearly in t, geometrically in delta), angular_distribution, or
to do full 3D interpolation at a t, delta, theta point.